refactor,test: 패키지 분리
