+++ src/main/java/step2/model/LottoTicket.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.model;
 
 import java.util.List;
