+++ src/main/java/step2/model/LottoTickets.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.model;
 
 import java.util.List;
