+++ src/test/java/step2/model/LottoNumberGeneratorTest.java
@@ -1,7 +1,9 @@
-package step2;
+package step2.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.model.LottoNumber;
+import step2.model.LottoNumberGenerator;
 
 import java.util.List;
