+++ src/test/java/step2/model/LottoTicketsTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.model;
 
 import org.junit.jupiter.api.Test;
