+++ src/test/java/step2/model/MoneyTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,6 +6,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import step2.model.Money;
 
 import java.util.stream.Stream;
